Build OneSignalFramework without dependency on location and IAM
The public protocols for OSInAppMessages and OSLocation have been moved to OneSignalCore along with a stub implementation.
This is a PoC implementation using ClassFromString but we could use injection instead to get the concrete implementation of OneSignalLocation and OneSignalInAppMessages without using classfromstring
