+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/OneSignalInAppMessages.h
@@ -26,113 +26,7 @@
  */
 
 #import <Foundation/Foundation.h>
-
-@interface OSInAppMessage : NSObject
-
-@property (strong, nonatomic, nonnull) NSString *messageId;
-
-// Convert the object into a NSDictionary
-- (NSDictionary *_Nonnull)jsonRepresentation;
-
-@end
-
-
-@interface OSInAppMessageTag : NSObject
-
-@property (strong, nonatomic, nullable) NSDictionary *tagsToAdd;
-@property (strong, nonatomic, nullable) NSArray *tagsToRemove;
-
-// Convert the class into a NSDictionary
-- (NSDictionary *_Nonnull)jsonRepresentation;
-
-@end
-
-typedef NS_ENUM(NSUInteger, OSInAppMessageActionUrlType) {
-    OSInAppMessageActionUrlTypeSafari,
-    
-    OSInAppMessageActionUrlTypeWebview,
-    
-    OSInAppMessageActionUrlTypeReplaceContent
-};
-
-@interface OSInAppMessageClickResult : NSObject
-
-// The action name attached to the IAM action
-@property (strong, nonatomic, nullable) NSString *actionId;
-
-// The URL (if any) that should be opened when the action occurs
-@property (strong, nonatomic, nullable) NSString *url;
-
-// Whether or not the click action dismisses the message
-@property (nonatomic) BOOL closingMessage;
-
-// Determines where the URL is loaded, ie. app opens a webview
-@property (nonatomic) OSInAppMessageActionUrlType urlTarget;
-
-// Convert the class into a NSDictionary
-- (NSDictionary *_Nonnull)jsonRepresentation;
-
-@end
-
-@interface OSInAppMessageWillDisplayEvent : NSObject
-@property (nonatomic, readonly, nonnull) OSInAppMessage *message;
-@end
-
-@interface OSInAppMessageDidDisplayEvent : NSObject
-@property (nonatomic, readonly, nonnull) OSInAppMessage *message;
-@end
-
-@interface OSInAppMessageWillDismissEvent : NSObject
-@property (nonatomic, readonly, nonnull) OSInAppMessage *message;
-@end
-
-@interface OSInAppMessageDidDismissEvent : NSObject
-@property (nonatomic, readonly, nonnull) OSInAppMessage *message;
-@end
-
-@protocol OSInAppMessageLifecycleListener <NSObject>
-@optional
-- (void)onWillDisplayInAppMessage:(OSInAppMessageWillDisplayEvent *_Nonnull)event
-NS_SWIFT_NAME(onWillDisplay(event:));
-- (void)onDidDisplayInAppMessage:(OSInAppMessageDidDisplayEvent *_Nonnull)event
-NS_SWIFT_NAME(onDidDisplay(event:));
-- (void)onWillDismissInAppMessage:(OSInAppMessageWillDismissEvent *_Nonnull)event
-NS_SWIFT_NAME(onWillDismiss(event:));
-- (void)onDidDismissInAppMessage:(OSInAppMessageDidDismissEvent *_Nonnull)event
-NS_SWIFT_NAME(onDidDismiss(event:));
-@end
-
-@interface OSInAppMessageClickEvent : NSObject
-@property (nonatomic, readonly, nonnull) OSInAppMessage *message;
-@property (nonatomic, readonly, nonnull) OSInAppMessageClickResult *result;
-// Convert the class into a NSDictionary
-- (NSDictionary *_Nonnull)jsonRepresentation;
-@end
-
-@protocol OSInAppMessageClickListener <NSObject>
-- (void)onClickInAppMessage:(OSInAppMessageClickEvent *_Nonnull)event
-NS_SWIFT_NAME(onClick(event:));
-@end
-
-/**
- Public API for the InAppMessages namespace.
- */
-@protocol OSInAppMessages <NSObject>
-
-+ (void)addTrigger:(NSString * _Nonnull)key withValue:(NSString * _Nonnull)value;
-+ (void)addTriggers:(NSDictionary<NSString *, NSString *> * _Nonnull)triggers;
-+ (void)removeTrigger:(NSString * _Nonnull)key;
-+ (void)removeTriggers:(NSArray<NSString *> * _Nonnull)keys;
-+ (void)clearTriggers;
-// Allows Swift users to: OneSignal.InAppMessages.paused = true
-+ (BOOL)paused NS_REFINED_FOR_SWIFT;
-+ (void)paused:(BOOL)pause NS_REFINED_FOR_SWIFT;
-
-+ (void)addClickListener:(NSObject<OSInAppMessageClickListener> *_Nullable)listener NS_REFINED_FOR_SWIFT;
-+ (void)removeClickListener:(NSObject<OSInAppMessageClickListener> *_Nullable)listener NS_REFINED_FOR_SWIFT;
-+ (void)addLifecycleListener:(NSObject<OSInAppMessageLifecycleListener> *_Nullable)listener NS_REFINED_FOR_SWIFT;
-+ (void)removeLifecycleListener:(NSObject<OSInAppMessageLifecycleListener> *_Nullable)listener NS_REFINED_FOR_SWIFT;
-@end
+#import <OneSignalCore/OSInAppMessages.h>
 
 @interface OneSignalInAppMessages : NSObject <OSInAppMessages>
