+++ iOS_SDK/OneSignalSDK/OneSignalLocation/OneSignalLocation.h
@@ -27,6 +27,8 @@
 
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
+#import <OneSignalCore/OSLocation.h>
+
 
 #ifndef OneSignalLocation_h
 #define OneSignalLocation_h
@@ -41,13 +43,6 @@ typedef struct os_last_location {
     double verticalAccuracy;
     double horizontalAccuracy;
 } os_last_location;
-
-@protocol OSLocation <NSObject>
-// - Request and track user's location
-+ (void)requestPermission;
-+ (void)setShared:(BOOL)enable NS_REFINED_FOR_SWIFT;
-+ (BOOL)isShared NS_REFINED_FOR_SWIFT;
-@end
 
 @interface OneSignalLocation : NSObject<OSLocation>
 + (Class<OSLocation>)Location;
